refactored tools/utils into common folder
