+++ src/jadhav_lab_to_nwb/olson_2024/olson_2024_behavior_interface.py
@@ -11,7 +11,7 @@ from neuroconv.utils import get_base_schema
 from pynwb.behavior import BehavioralEvents
 from pynwb import TimeSeries
 
-from .tools.spikegadgets import readTrodesExtractedDataFile
+from ..common.tools.spikegadgets import readTrodesExtractedDataFile
 
 
 class Olson2024BehaviorInterface(BaseDataInterface):
